Version 0.5 of the Noiz
This version is the last version I was working at while working for
Unistra. This version is compatible with the report finalizing my
project and the data I produced. If you want to use that data, use this
version.

The only difference, apart of minor CI compatibility changes,
is anonimization of the test dataset.

Test data that are used for running system tests were changed and old
files were removed from the history of the repository. This means that,
while full development history of the code was preserved, running system
tests for commits from the past will likely fail. There might be also
some unit tests that will fail.

Damian Kula
heavelock@gmail.com

Signed-off-by: Damian Kula <heavelock@gmail.com>